Place a filter-model on top of the list-model
This changes the old model-hierarchy from

List <-> Filter <-> Tree <-> Core

to

Filter <-> List <-> Tree <-> Core

The idea is that since QSortFilterProxyModel is so ineffective,
let's give it less data to process. Since only one trip is expanded
in the list-model, it has significantly less data than the tree
model.

Is this a good idea? I don't know!

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>